update job params list
