+++ .env.example
@@ -30,7 +30,6 @@ PREFECT_API_URL=http://prefect:4200/api
 FLOW_NAME="Parent flow/launch_parent_flow"
 TIMEZONE="US/Pacific"
 PREFECT_TAGS='["latent-space-explorer"]'
-FLOW_TYPE="docker"
 
 # MLFlow
 MLFLOW_TRACKING_URI=http://mlflow:5000
