+++ docker-compose.yml
@@ -135,7 +135,6 @@ services:
       FLOW_NAME: '${FLOW_NAME}'
       TIMEZONE: "${TIMEZONE}"
       PREFECT_TAGS: "${PREFECT_TAGS}"
-      FLOW_TYPE: "${FLOW_TYPE}"
       CONTAINER_NETWORK: "${CONTAINER_NETWORK}"
       # Slurm jobs
       PARTITIONS_CPU: "${PARTITIONS_CPU}"
