+++ src/utils/job_utils.py
@@ -13,17 +13,6 @@ MLFLOW_TRACKING_URI = os.getenv("MLFLOW_TRACKING_URI", "")
 MLFLOW_TRACKING_USERNAME = os.getenv("MLFLOW_TRACKING_USERNAME", "")
 MLFLOW_TRACKING_PASSWORD = os.getenv("MLFLOW_TRACKING_PASSWORD", "")
 
-# Flow parameters
-PARTITIONS_CPU = json.loads(os.getenv("PARTITIONS_CPU", "[]"))
-RESERVATIONS_CPU = json.loads(os.getenv("RESERVATIONS_CPU", "[]"))
-MAX_TIME_CPU = os.getenv("MAX_TIME_CPU", "1:00:00")
-PARTITIONS_GPU = json.loads(os.getenv("PARTITIONS_CPU", "[]"))
-RESERVATIONS_GPU = json.loads(os.getenv("RESERVATIONS_CPU", "[]"))
-MAX_TIME_GPU = os.getenv("MAX_TIME_CPU", "1:00:00")
-SUBMISSION_SSH_KEY = os.getenv("SUBMISSION_SSH_KEY", "")
-FORWARD_PORTS = json.loads(os.getenv("FORWARD_PORTS", "[]"))
-CONTAINER_NETWORK = os.getenv("CONTAINER_NETWORK", "")
-
 
 def parse_tiled_url(url, user, project_name, tiled_base_path="/api/v1/metadata"):
     """
@@ -42,14 +31,12 @@ def parse_train_job_params(
     model_parameters,
     user,
     project_name,
-    flow_type,
     latent_space_params,
     dim_reduction_params,
 ):
     """
     Parse training job parameters
     """
-    # TODO: Use model_name to define the conda_env/algorithm to be executed
     data_uris = [dataset.uri for dataset in data_project.datasets]
 
     results_dir = f"{WRITE_DIR}/{user}"
@@ -69,150 +56,42 @@ def parse_train_job_params(
         "mlflow_tracking_password": MLFLOW_TRACKING_PASSWORD,
     }
 
-    ls_python_file_name_train = latent_space_params["python_file_name"]["train"]
-    ls_python_file_name_inference = latent_space_params["python_file_name"]["inference"]
-    dm_python_file_name = dim_reduction_params["python_file_name"]
+    # Create a simpler params_list structure with model_name and task_name
+    params_list = [
+        {
+            "model_name": latent_space_params["model_name"],
+            "task_name": "train",
+            "params": {
+                "io_parameters": io_parameters,
+                "model_parameters": model_parameters,
+            },
+        },
+        {
+            "model_name": latent_space_params["model_name"],
+            "task_name": "inference",
+            "params": {
+                "io_parameters": io_parameters,
+                "model_parameters": model_parameters,
+            },
+        },
+        {
+            "model_name": dim_reduction_params["model_name"],
+            "task_name": "run",
+            "params": {
+                "io_parameters": io_parameters,
+                "model_parameters": {
+                    "n_components": 2,
+                    "min_dist": 0.1,
+                    "n_neighbors": 5,
+                },
+            },
+        },
+    ]
 
-    if flow_type == "podman" or "docker":
-        job_params = {
-            "flow_type": flow_type,
-            "params_list": [
-                {
-                    "image_name": latent_space_params["image_name"],
-                    "image_tag": latent_space_params["image_tag"],
-                    "command": f"python {ls_python_file_name_train}",
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": model_parameters,
-                    },
-                    "volumes": [
-                        f"{READ_DIR_MOUNT}:/tiled_storage",
-                    ],
-                    "network": CONTAINER_NETWORK,
-                },
-                {
-                    "image_name": latent_space_params["image_name"],
-                    "image_tag": latent_space_params["image_tag"],
-                    "command": f"python {ls_python_file_name_inference}",
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": model_parameters,
-                    },
-                    "volumes": [
-                        f"{READ_DIR_MOUNT}:/tiled_storage",
-                    ],
-                    "network": CONTAINER_NETWORK,
-                },
-                {
-                    "image_name": dim_reduction_params["image_name"],
-                    "image_tag": dim_reduction_params["image_tag"],
-                    "command": f"python {dm_python_file_name}",
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": {
-                            "n_components": 2,
-                            "min_dist": 0.1,
-                            "n_neighbors": 5,
-                        },
-                    },
-                    "volumes": [
-                        f"{READ_DIR_MOUNT}:/tiled_storage",
-                    ],
-                    "network": CONTAINER_NETWORK,
-                },
-            ],
-        }
-
-    elif flow_type == "conda":
-        job_params = {
-            "flow_type": "conda",
-            "params_list": [
-                {
-                    "conda_env_name": latent_space_params["conda_env"],
-                    "python_file_name": ls_python_file_name_train,
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": model_parameters,
-                    },
-                },
-                {
-                    "conda_env_name": latent_space_params["conda_env"],
-                    "python_file_name": ls_python_file_name_inference,
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": model_parameters,
-                    },
-                },
-                {
-                    "conda_env_name": dim_reduction_params["conda_env"],
-                    "python_file_name": dm_python_file_name,
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": {
-                            "n_components": 2,
-                            "min_dist": 0.1,
-                            "n_neighbors": 5,
-                        },
-                    },
-                },
-            ],
-        }
-
-    else:
-        job_params = {
-            "flow_type": "slurm",
-            "params_list": [
-                {
-                    "job_name": "latent_space_explorer",
-                    "num_nodes": 1,
-                    "partitions": PARTITIONS_CPU,
-                    "reservations": RESERVATIONS_CPU,
-                    "max_time": MAX_TIME_CPU,
-                    "conda_env_name": latent_space_params["conda_env"],
-                    "python_file_name": ls_python_file_name_train,
-                    "submission_ssh_key": SUBMISSION_SSH_KEY,
-                    "forward_ports": FORWARD_PORTS,
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": model_parameters,
-                    },
-                },
-                {
-                    "job_name": "latent_space_explorer",
-                    "num_nodes": 1,
-                    "partitions": PARTITIONS_CPU,
-                    "reservations": RESERVATIONS_CPU,
-                    "max_time": MAX_TIME_CPU,
-                    "conda_env_name": latent_space_params["conda_env"],
-                    "python_file_name": ls_python_file_name_inference,
-                    "submission_ssh_key": SUBMISSION_SSH_KEY,
-                    "forward_ports": FORWARD_PORTS,
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": model_parameters,
-                    },
-                },
-                {
-                    "job_name": "latent_space_explorer",
-                    "num_nodes": 1,
-                    "partitions": PARTITIONS_CPU,
-                    "reservations": RESERVATIONS_CPU,
-                    "max_time": MAX_TIME_CPU,
-                    "conda_env_name": dim_reduction_params["conda_env"],
-                    "python_file_name": dm_python_file_name,
-                    "submission_ssh_key": SUBMISSION_SSH_KEY,
-                    "forward_ports": FORWARD_PORTS,
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": {
-                            "n_components": 2,
-                            "min_dist": 0.1,
-                            "n_neighbors": 5,
-                        },
-                    },
-                },
-            ],
-        }
+    # Keep the job params simplified
+    job_params = {
+        "params_list": params_list,
+    }
 
     return job_params
 
@@ -222,14 +101,12 @@ def parse_inference_job_params(
     model_parameters,
     user,
     project_name,
-    flow_type,
     latent_space_params,
     dim_reduction_params,
 ):
     """
     Parse inference job parameters
     """
-    # TODO: Use model_name to define the conda_env/algorithm to be executed
     data_uris = [dataset.uri for dataset in data_project.datasets]
 
     results_dir = f"{WRITE_DIR}/{user}"
@@ -249,118 +126,34 @@ def parse_inference_job_params(
         "mlflow_tracking_password": MLFLOW_TRACKING_PASSWORD,
     }
 
-    ls_python_file_name_inference = latent_space_params["python_file_name"]["inference"]
-    dm_python_file_name = dim_reduction_params["python_file_name"]
+    # Create a simpler params_list structure with model_name and task_name
+    params_list = [
+        {
+            "model_name": latent_space_params["model_name"],
+            "task_name": "inference",
+            "params": {
+                "io_parameters": io_parameters,
+                "model_parameters": model_parameters,
+            },
+        },
+        {
+            "model_name": dim_reduction_params["model_name"],
+            "task_name": "run",
+            "params": {
+                "io_parameters": copy.copy(io_parameters),  # Ensures uid_retrieve is empty
+                "model_parameters": {
+                    "n_components": 2,
+                    "min_dist": 0.1,
+                    "n_neighbors": 5,
+                },
+            },
+        },
+    ]
 
-    if flow_type == "podman" or "docker":
-        job_params = {
-            "flow_type": flow_type,
-            "params_list": [
-                {
-                    "image_name": latent_space_params["image_name"],
-                    "image_tag": latent_space_params["image_tag"],
-                    "command": f"python {ls_python_file_name_inference}",
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": model_parameters,  # Default parameters
-                    },
-                    "volumes": [
-                        f"{READ_DIR_MOUNT}:/tiled_storage",
-                    ],
-                    "network": CONTAINER_NETWORK,
-                },
-                {
-                    "image_name": dim_reduction_params["image_name"],
-                    "image_tag": dim_reduction_params["image_tag"],
-                    "command": f"python {dm_python_file_name}",
-                    "params": {
-                        "io_parameters": copy.copy(
-                            io_parameters
-                        ),  # Ensures uid_retrieve is empty
-                        "model_parameters": {
-                            "n_components": 2,
-                            "min_dist": 0.1,
-                            "n_neighbors": 5,
-                        },
-                    },
-                    "volumes": [
-                        f"{READ_DIR_MOUNT}:/tiled_storage",
-                    ],
-                    "network": CONTAINER_NETWORK,
-                },
-            ],
-        }
-    elif flow_type == "conda":
-        job_params = {
-            "flow_type": "conda",
-            "params_list": [
-                {
-                    "conda_env_name": latent_space_params["conda_env"],
-                    "python_file_name": ls_python_file_name_inference,
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": model_parameters,
-                    },
-                },
-                {
-                    "conda_env_name": dim_reduction_params["conda_env"],
-                    "python_file_name": dm_python_file_name,
-                    "params": {
-                        "io_parameters": copy.copy(
-                            io_parameters
-                        ),  # Ensures uid_retrieve is empty
-                        "model_parameters": {
-                            "n_components": 2,
-                            "min_dist": 0.1,
-                            "n_neighbors": 5,
-                        },
-                    },
-                },
-            ],
-        }
-
-    else:
-        job_params = {
-            "flow_type": "slurm",
-            "params_list": [
-                {
-                    "job_name": "latent_space_explorer",
-                    "num_nodes": 1,
-                    "partitions": PARTITIONS_CPU,
-                    "reservations": RESERVATIONS_CPU,
-                    "max_time": MAX_TIME_CPU,
-                    "conda_env_name": latent_space_params["conda_env"],
-                    "python_file_name": ls_python_file_name_inference,
-                    "submission_ssh_key": SUBMISSION_SSH_KEY,
-                    "forward_ports": FORWARD_PORTS,
-                    "params": {
-                        "io_parameters": io_parameters,
-                        "model_parameters": model_parameters,
-                    },
-                },
-                {
-                    "job_name": "latent_space_explorer",
-                    "num_nodes": 1,
-                    "partitions": PARTITIONS_CPU,
-                    "reservations": RESERVATIONS_CPU,
-                    "max_time": MAX_TIME_CPU,
-                    "conda_env_name": dim_reduction_params["conda_env"],
-                    "python_file_name": dm_python_file_name,
-                    "submission_ssh_key": SUBMISSION_SSH_KEY,
-                    "forward_ports": FORWARD_PORTS,
-                    "params": {
-                        "io_parameters": copy.copy(
-                            io_parameters
-                        ),  # Ensures uid_retrieve is empty
-                        "model_parameters": {
-                            "n_components": 2,
-                            "min_dist": 0.1,
-                            "n_neighbors": 5,
-                        },
-                    },
-                },
-            ],
-        }
+    # Keep the job params simplified
+    job_params = {
+        "params_list": params_list,
+    }
 
     return job_params
 
@@ -376,7 +169,7 @@ def parse_model_params(model_parameters_html, log, percentiles, mask):
         # param["props"]["children"][0] is the label
         # param["props"]["children"][1] is the input
         parameter_container = param["props"]["children"][1]
-        # The achtual parameter item is the first and only child of the parameter container
+        # The actual parameter item is the first and only child of the parameter container
         parameter_item = parameter_container["props"]["children"]["props"]
         key = parameter_item["id"]["param_key"]
         if "value" in parameter_item:
